Implemented map pin to be clickable. The on click event is now just a toast message.
Updated the map pin calling in main activity

+++ MasterProject/MuseumApp/app/src/main/java/com/example/museumapp/ShowCase.java
@@ -6,11 +6,11 @@ public class ShowCase {
     private int closetID;
     private float length;
     private float width;
-    private int x, y;
+    private float x, y;
     private int floorNum;
     private List<Item> items;
 
-    public ShowCase(int closetID, float length, float width, int x, int y, int floorNum, List<Item> items) {
+    public ShowCase(int closetID, float length, float width, float x, float y, int floorNum, List<Item> items) {
         this.closetID = closetID;
         this.length = length;
         this.width = width;
@@ -44,7 +44,7 @@ public class ShowCase {
         this.width = width;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
@@ -52,7 +52,7 @@ public class ShowCase {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
